requirements.txt
langchain==0.3.1
langchain-groq==0.1.3
langchain-community==0.3.1
gradio==4.39.0
python-dotenv==1.0.1
tavily-python==0.3.3
